Multiple changes.
+ Improved workflow
+ Changes to extract_mcmc_draws and project_using_fitted to avoid
reading in incidence file. This is because I had not fixed the issue
with incorrect incidence in Ghana before running the model. So
I ran the model from days 21 to 63 with the corrected file. I want to
ensure that the projection uses the incidence that was used to fit.
+ A new file for creating files for the manuscript.
+ updated .gitignore
+ .lab-notebook is a list of all non-reproducible changes made to
files. That is, every time I had to change a file by hand.